Remove the space before exclamation points
I have no idea why this is a thing, but it's consistent with the entire project so I assume it to be squirt's preferred way of punctuation. It's not mine though! I'll leave it in the comments, but in any place where a player can see it, I'm changing it.