Remove the speed limit from grinding
If you move too slow on a grind, you will fail it. The problem is that this limit is very arbitrary, I could not find a value change which made sense. Lowering it barely makes a difference. The downside of this is removing the usefulness of momentum shifted multi grinds, but this is a relatively unknown tech. It isn't even documented in the wiki, and is useless due to the removal of stamina.