Support for dynamic and faster grinds
Adds support for dynamic grind speeds. If the player is going slower than the default speed, they will be given the default speed (identical behavior to the original Kickflip Quake.) The default speed has been changed to be 40 units faster. If the player is going faster than the default speed, that speed will carry over to the grind. Over time this speed will be lost.