Halved Z change in speed.
odd to explain, but when on a grind you are constantly pushed down at a rate of 280 units. Changing this fixes the featherfalling issue which happens when sliding off of a rail, but also halves the loss / gain in speed which occurs when going up / down a rail. Might be worth it, not sure