Adds new variables and better comments
Added comments where I think they'd be important. I also added the "viewsize" (changes the visibility of the HUD, default is OFF) and the "gl_font" (changes the HUD and menu font) variables. I don't play with the HUD on by default, and I find it annoying to constantly change my fonts when I clone KFQuake.