Removes gl_font, moves autoexec, removes f_startup
"gl_font" has to be removed in order for it to work in the autoexec.cfg, and moves the "exec Autoexec.cfg" to the bottom of the file in order to ensure nothing overwrites it. We also removed the "f_startup" section because it doesn't do anything useful. I don't like it.