Having penta makes all tricks have a boost jump.
The coolness factor of the trick is what determines how far it sends you (ex. a Kickflip sends you 400 units up, a 360 Shove-it will send you 700 units up, and a Backside 360 will send you a staggering 1300 units up.) That is very rarely practical, but it's really funny! Not sure what I want to do with the ring of invis, probably will not give it an ability. Maybe I'll get it one later, but I kinda doubt it. The reason being that I don't like how this powerup makes you look, and I don't even know what ability I could give it; especially because I want all powerups to be able to work with eachother.