Rebinds the Backside 360 to a key instead of combo
There is an incredbily annoying glitch that occurs when doing a Backside 360 in Kickflip Quake. Oftentimes, pressing the trick1 and trick2 keys will almost always cause too many impulse commands. This makes it so impulse 98 is ignored, which means trick1 will be held down forever until you press Shift again. This means that everytime you use Backside 360, you HAVE to press Shift again just to not risk getting sticky keys. This is dumb, and it's more comfortable to have the Backside 360 bound to a key anyway.